+++ src/TraCIAPI/Utils.h
@@ -62,8 +62,8 @@ public:
 uint32_t RGB2HEX(uint8_t r, uint8_t g, uint8_t b);
 void HEX2RGB(uint32_t hex, uint8_t& r, uint8_t& g, uint8_t& b);
 
-int MPH2MS(int mph);
-int MS2MPH(int ms);
+double MPH2MS(double mph);
+double MS2MPH(double ms);
 
 namespace traci_api
 {
